Pipeline lints Markdown files.
This ensures that all Markdown documents in this repo, including
README.md and CONTRIBUTING.md follow Beautiful Canoe standards.
See the .mdl.rb file for details.

Note that the lint:markdown part of the pipeline is run on GitLab
shared runners, whereas the rest of the pipeline is run on our own
servers.

Part-of: <https://gitlab.com/traffic3d/traffic3d/-/merge_requests/88>