Use GitLab rules syntax.
GitLab is deprecating only/except syntax, in favour of the new grammar for rules.
This commit moves us to the new syntax. It should NOT change the semantics of the CI config.

The deprecation notice is here:

https://about.gitlab.com/releases/2020/04/22/gitlab-12-10-released/#gitlab-ci/cd-pipeline-configuration-templates-are-transitioning-to-%60rules%60-instead-of-%60only/except%60

and the relevant issues are:

* https://gitlab.com/gitlab-org/gitlab/-/issues/213336
* https://gitlab.com/groups/gitlab-org/-/epics/2300

Tested-by: Bob The Beautiful Canoe Robot <https://gitlab.com/traffic3d/traffic3d/-/merge_requests/96>
Part-of: <https://gitlab.com/traffic3d/traffic3d/-/merge_requests/96>