CI config uses build matrix syntax
This commit introductes parallel matrix jobs to the CI config:

https://docs.gitlab.com/ee/ci/yaml/README.html#parallel-matrix-jobs

The goal here is not to change the behaviour of the pipeline, but
to simplify the job descriptions, so that we only define one job
for the build stage of the pipeline.

This is in preparation for later commits, which will specialse
the build job for different branches.

Part-of: <https://gitlab.com/traffic3d/traffic3d/-/merge_requests/120>