Check for unecessary .meta files in CI
This commit adds a new lint in a file ./ci/metalint.sh which
checks for .meta files that do not have a corresponding asset.

This has previously caused a pipeline failure in the test
stage, but shifting the check to the start of the pipeline
means that pipelines can fail fast.

Part-of: <https://gitlab.com/traffic3d/traffic3d/-/merge_requests/141>