Refactors evacuAgent behaviour to strategies
Note - This change removes default behaviour for non-shooter
agents which causes exceptions to be thrown.

This is necessary to ensure that behaviour in the pedestrian
update method does not interfere with evacuAgent behaviour.

Adds the following folders:

BehaviourStructures - To store behaviour sturecture code like
behaviourCollection and behaviourStrategy.

BehaviourTypes - To store enums of different behaviour sets.

Factories - To store factory patterns to create and add
behaviourStrategies to agents.

ShooterBehaviours - To store the inidividuasl behaviour scripts
for shooter sgent types.

UI - To store scripts associated with the UI.

Adds the following scripts:

BehaviourCollection - Stores and iterates over a collection of
behaviourStrategies and is stored as a child of a behaviourController.

BehaviourCollectionStrategy - Defines the standard format of
BehaviourCollections.

BehaviourController - Stores a collection of BehaviourCollections
and iterates over the current behaviourCollection state to perform
behaviours.

BehaviourStrategy - Defines the format of a single behaviour with
methods for detecting if the condtition of the behaviour is met
and for performing the behaviour.

BehaviourTypeAttribute - Defines a custom attribute for storing the
name and chance of a given behaviourStrategy.

ShooterBehaviourType - An enum to store the preference order of
behaviourStrategies with an associated behaviourTypeAttribute on each.

BehaviourCollectionFactory - Creates a behaviourCollection based on
a behaviourType enum paramater, which is added as a child of
a behaviourController.

BehaviourCollectionFactoryHelper - A static helper class to parse string
names stored in behaviourTypeAttributes as types for instantiation.

PedestrianBehaviourFactory - Adds all evacuation behaviour and necessary
child objects and tags to pedestrians.

Part-of: <https://gitlab.com/traffic3d/traffic3d/-/merge_requests/142>