Fixes and refactors FieldOfViewTests
Removes the pedestrians from FieldOfViewTests in favour of
the fieldOfView prefab and primiteves with necessary transforms
and layers.

Change needed due to the different GameObject hierarchy and inability
to stop BehaviourController performing behaviours even when none are
there such as in the case of unit tests.

Part-of: <https://gitlab.com/traffic3d/traffic3d/-/merge_requests/142>