Fixes InformationPanelControllerTests
As part of the refactor adding BehaviourControllers to Pedestrians
it is not a good strategy to spawn entire pedestrians for testing.

Instead empty GameObjects with the correct tags and inactive
Pedestrian scripts are spawned.

Also removes the reliance on the Scene by instantiating a
UserInterface prefab so that this logic can be tested in isolation
as before it looked through the scene for a UserInterface.

Part-of: <https://gitlab.com/traffic3d/traffic3d/-/merge_requests/169>