Changes in line with comments
FieldOfView is now a child of the Pedestrian GameObject.

Refactors BehaviourTypes and BehaviourTypeOrder from Enum into class
based version.

Removes:

* BehaviourTypeAttribute.
* ShooterBehaviourTypes enum.
* BehaviourCollectionFactoryHelper.

Adds:

* BehaviourType, an abstract class responsible for holding the
name of a script and the chancce it has of being added.
* BehaviourTypeOrder, an abstract class responsible foe holding and
returning a list of BehaviourType.
* ShooterBehaviourTypeOrder, takes over job of ShooterBehaviourTypes
enum by maintaining an orderded collection of behaviours that are
added to a BehaviourCollection.
* Tests for BehaviourCollectionFactory.

Also adds condition to BehaviourController.FixedUpdate() to stop
behaviours being performed, which leads to exceptions in tests.

Tested-by: Bob The Beautiful Canoe Robot <https://gitlab.com/traffic3d/traffic3d/-/merge_requests/142>
Part-of: <https://gitlab.com/traffic3d/traffic3d/-/merge_requests/142>