Refactors pedestrianTypes into classes
Note tests have not been fixed in this commit.

Removes BehaviourCollectionFactory and refactors
PedestrianBehaviourFactory into AbstractEvacuAgentPedestrianFactory
which is the base class for factories to produce pedestrianTypes.

PedestrianTypes are now class based with base class
EvacuAgentPedestrianBase.

ShooterPedestrian and WorkerPedestrian model each respective
pedestrianType in terms of their highlighter and
PedestrianPointPathCreator, each is now a prefab.

Adds new hierarchy to PedestrianPointPathCreator with  abstract base
class PedestrianPointPathCreator:

* ShooterPedestrianPointPathCreator derives from PedestrianPointPathCreator
to create paths for shooter agents

* NonShooterPedestrianPointPathCreator derives from PedestrianPointPathCreator
and has further children like WorkerPedestrianPointPathCreator
that are used to create paths based on PedestrianType.

Adds AddEvacuAgentBehaviour() to PedestrianFactory which
calls the appropriate factory to create a pedestrianType, this
is based on a counter of each pedestrian type.

Part-of: <https://gitlab.com/traffic3d/traffic3d/-/merge_requests/169>