Refactors PedestrianPointPathCreator to Vector3
PedestrianPointPathCreator.CreatePath() was used to create paths
of PedestrianPoint, this is now changed to create paths of Vector3
to allow points on the NavMesh to be added to path creation.

Also fixes the tests this breaks, tests are very similar and now
use EvacuAgentCommonSceneTest.GetPedestrianPointFromLocation() to get
PedestrianPoint from their Vector3 locations.

Also adds initial changes for FriendGroupLeaderType via the
FriendGroupLeaderTypeOrder and LeaderFollowerPedestrianFactor.

Tested-by: Bob The Beautiful Canoe Robot <https://gitlab.com/traffic3d/traffic3d/-/merge_requests/160>
Part-of: <https://gitlab.com/traffic3d/traffic3d/-/merge_requests/160>