Adds initial structure for group movement
Note: This is still under construction and this commit is being made
at an experimental point to decrease the size of the commits for this
MR.

As part of the experimental nature some code is commented out so that
boid behaviours can be isolated. Most notably factories are
commented out in PedestrianFactory and some BehaviourTypeOrders are
also commented out of FriendGroupFollowerTypeOrder and
FriendGroupLeaderTypeOrder.

This Commit adds:

The Boids folder containing:

* Boid Components: A folder for storing components used to calculate
velocity vectors for group dynamics behaviours.

* BoidBehaviourStrategyBase: An abstract base class for common boid
data that will be used in the BehaviourStrategy hierarchy for all
boid based pedestrians.

* FriendGroupBoidBehaviour: A child class of BoidBehaviourStrategyBase
that will be used by FiendGroupFollower pedestrian types. This class
stores the appropriate BoidComponent weights and uses BoidManager
to calculate new velocities.

* BoidManager: Iterates over all children of BoidComponentBase
attached to the GameObject. Both the manager and components are
added to the appropriate pedestrian prefab.

* GroupCollection: A class used to store information about a group
of pedestrians used by boids based pedestrian types.

This commit changes:

* FieldOfView view angle and radius for use in boids based pedestrians,
this is necessary as a narrow cone of vision can lead to single file
following patterns.

Part-of: <https://gitlab.com/traffic3d/traffic3d/-/merge_requests/165>