Moves to boid based system from NavMeshAgent
The majority of movement is now calculated from boid based velocities.

These velocities are a sum of multiple boid components which are then
passed to the NavMeshAgent.

Paramaters for these components have been found through rounds of play
testing to see which produces the most realistic movement.

As part of this commit FieldOfView has been updated to detect nearby
obstacles to be used in the obstacle avoidance componenet. This is
necessary as the boids based system broke the use of the NavMeshAgent
for obstacle avoidance.

To detect a main ray cast is made forward, and several peripheral rays
are cast to the sides and diagonally in front of the agent. The
peripheral rays are shorter and produce less force on the pedestrian.
These are mainly used to navigate corners where the ahead ray is not
intersecting an obstacle.

Part-of: <https://gitlab.com/traffic3d/traffic3d/-/merge_requests/165>