Remove ts from CI build and test
We used ts to time the operations in our test and build stages.
This has only rarely been used (when we were trying to improve
pipeline speeds) and isn't needed for day to day development.

If we do need it again, it would make sense to install moreutils
in the pipeline, and then once the relevant MR has been reviewed,
remove it before merging.

Part-of: <https://gitlab.com/traffic3d/traffic3d/-/merge_requests/150>