Add editorconfig with BC default style
This encodes our standards for indentation, trailing whitespace
and so on. See: https://editorconfig.org/ for more documentation.

The editorconfig page also has a list of IDEs that work with
editorconfig out of the box, and those that require plugins.

Part-of: <https://gitlab.com/traffic3d/traffic3d/-/merge_requests/152>