Add standard Python .gitignore
This is a standard Python .gitignore, minus some configuration
for web frameworks.

This configuration could have gone into the backend/ directory,
but we leave it here so that developers are free to create new
Python code anywhere in the repo, and configure the "root" of
their source tree (in their IDEs) wherever they wish.

Part-of: <https://gitlab.com/traffic3d/traffic3d/-/merge_requests/152>