Fixes PedestrianTypeFactoryTests
This was due to the name of the script file being different than it was
in the script file, as the class name. Changing both to match fixed the
failing tests thrown by null pointer errors.

Changes in line with review comments

Tested-by: Bob The Beautiful Canoe Robot <https://gitlab.com/traffic3d/traffic3d/-/merge_requests/165>
Part-of: <https://gitlab.com/traffic3d/traffic3d/-/merge_requests/169>