Fixes WaitAtDestinationBehaviourTests
The issue was to do with a random value "RadiusToDestination" being
used in WaitAtDestinationBehaviourTests. This property has been changed
to public and is set to an appropriate value in the test to ensure that
the expected value is less than the value for RadiusToDestination.

Part-of: <https://gitlab.com/traffic3d/traffic3d/-/merge_requests/169>