Temporarily remove common image formats from LFS
Because we have not yet merged a fix for issue #215 we have had some
problems with MRs which contain images, such as !171. Please see
issue #216 for more details.

We believe some of this may be due to tooling, and some of it may
be due to developers having different versions of git-lfs, but having
LFS defined in .gitattributes without having the whole repo migrated is
clearly going to lead to problems.

At this point, we do not want to resolve #215 because we are very close
to the next release, and will need more time to deal with any issues
that might arise from migrating the repo.

So, this commit remove images from LFS.

Tested-by: Bob The Beautiful Canoe Robot <https://gitlab.com/traffic3d/traffic3d/-/merge_requests/173>
Part-of: <https://gitlab.com/traffic3d/traffic3d/-/merge_requests/173>